move filter to plugins folder
